updates (#5)
* updates

- update readme
- add more options to maintain the original behavior along the new features
- emit an event on drag
- put back the move handle on click
- move styles to its own file
- removed eveyrthing but the component
- reversed the zoom wheel direction “up-zoomIn / down-zoomOut”
- add zoom “min, max”
- add some animation for better UX
- hide the handle when dram mode is on to avoid confusion & a cleaner UI

* updated readme

* restored yarn lock

* Update README.md

* fixing md lint MD033/no-inline-html: Inline HTML [Element: br]

* updated usage section

* Update styles.scss

* add `hideHandleOnDrag`

* Update README.md

* fixed variable naming

* simpler hideHandle

* fixed issue with blocked handle

* attempt to fix CI

* attempt 2 to fix CI
