+++ README.md
@@ -2,19 +2,23 @@
 
 ## Purpose of this fork
 
-- allow drag & drop images to compare them
-- allow drag & drop of a single image to change only one side
-- allow zoom & pan of the comparison to see details
+**simple images loading** via drag & drop files into browser, will load images locally (no upload to any server)
 
-You can see the updated features of this fork on : https://image-compare.netlify.com/
+* drop 2 files
+* drop 1 file on one side to change only one side of the comparison
 
-And still check the original author website : https://marcincichocki.github.io/vue-image-compare/
+**better image comparison**  to see details :
 
-Bellow, the original readme :
+* allow click & drag
+* allow zoom
+
+You can see the updated features of this fork on : [https://image-compare.netlify.com](https://image-compare.netlify.com)
+
+And still check the original author website : [https://marcincichocki.github.io/vue-image-compare](https://marcincichocki.github.io/vue-image-compare)
 
 ## Requirements
 
-- [Vue.js](http://vuejs.org/) (^2.0.0)
+* [Vue.js](http://vuejs.org/) (^2.0.0)
 
 ## Installation
 
@@ -24,59 +28,100 @@ npm i vue-image-compare --save
 
 ## Usage
 
-```javascript
-import Vue from 'vue';
-import VueImageCompare from 'vue-image-compare';
+At Vue bootstrap :
 
-Vue.use(VueImageCompare);
+```javascript
+import Vue from 'vue'
+import imageCompare from 'vue-image-compare'
 
 new Vue({
-  data() {
+    components: {
+        imageCompare
+    },
+    data() {
+        return {
+            before: '/img/before.jpg',
+            after: '/img/after.jpg'
+        }
+    }
+}).$mount('#app')
+```
+
+or inside another component :
+
+```javascript
+import imageCompare from 'vue-image-compare'
+
+export default {
+   data() {
     return {
       before: '/img/before.jpg',
       after: '/img/after.jpg'
     }
+  },
+  components: {
+    imageCompare
   }
-}).$mount('#app');
+}
 ```
+
+Then you can use `image-compare` with a minimal init :
 
 ```html
 <div id="app">
-  <image-compare :before="before" :after="after"/>
+    <image-compare :before="before" :after="after"/>
 </div>
 ```
 
-If you are **not** using using es6, instead of importing add 
+or with some optionals params, like in [demo](https://github.com/Shuunen/vue-image-compare/tree/demo) :
 
 ```html
-<script src="/path/to/vue-image-compare.js"></script>
+<div id="app">
+    <image-compare :before="before" :after="after" full isZoomable isSwitchable isDraggable>
+      <i class="fa fa-angle-left" aria-hidden="true" slot="icon-left"></i>
+      <i class="fa fa-angle-right" aria-hidden="true" slot="icon-right"></i>
+    </image-compare>
+</div>
 ```
 
-just before closing body tag. 
+### Props
 
-## Props
+| Name           | Type      | Default                    | Description                 |
+| -------------- | --------- | -------------------------- | --------------------------- |
+| `before`       | `String`  | `undefined` (**required**) | path to the image *before*  |
+| `after`        | `String`  | `undefined` (**required**) | path to the image *after*   |
+| `full`         | `Boolean` | `false`                    | stretch images (1)          |
+| `padding`      | `Object`  | `{left: 0, right: 0}`      | left and right padding (2)  |
+| `hideAfter`    | `Boolean` | `false`                    | hide the after image        |
+| `zoom`         | `Object`  | `{min: 0.5, max: 2}`       | scale image by              |
+| `reset`        | `Boolean` | `false`                    | reset all to original       |
+| `isZoomable`   | `Boolean` | `false`                    | mouse wheel to zoom in/out  |
+| `isDraggable`  | `Boolean` | `false`                    | allow moving the comparison |
+| `isSwitchable` | `Boolean` | `false`                    | allow drag & drop           |
+| `hideHandle`   | `Boolean` | `false`                    | hide vertical handle bar    |
+| `labels`       | `Object`  | `{after: '', before: ''}`  | comparison labels           |
 
-| Name | Type | Description | Required | Default |
-| --- | --- | --- | --- | --- |
-| `before` | `String` | Path to the image image *before* change | Yes | `undefined` |
-| `after` | `String` | Path to the image image *after* change | Yes | `undefined` |
-| `full` | `Boolean` | Determines if images are stretched to fill parent element. Can be used with help of CSS `object-fit: cover` to create full page image comparison | No | `false` |
-| `padding` | `Object` | Set left and right "padding" in pixels, so handle can **not** reach edge of an image | No | `{ left: 0, right: 0 }` |
-| `hideAfter` | `Boolean` | Hide image after and handle | No | `false` |
+(1) : Determines if images are stretched to fill parent element. Can be used with help of CSS object-fit: cover to create full page image comparison
 
-## Slots
+(2) : Set left and right "padding" in pixels, so handle can not reach edge of an image
 
-- `icon-left` - element to be placed on the left side of the handle
-- `icon-right` - element to be placed on the right side of the handle
+### Slots
 
-Example: 
+* `icon-left` - element to be placed on the left side of the handle
+* `icon-right` - element to be placed on the right side of the handle
+
+Example:
 
 ```html
-<image-compare before="/img/before.jpg" after="/img/after.jpg" :padding="{ left: 50, right: 50 }">
-  <i class="fa fa-angle-left" aria-hidden="true" slot="icon-left"></i>
-  <i class="fa fa-angle-right" aria-hidden="true" slot="icon-right"></i>
+<image-compare before="/img/before.jpg" after="/img/after.jpg">
+    <i class="fa fa-angle-left" aria-hidden="true" slot="icon-left"></i>
+    <i class="fa fa-angle-right" aria-hidden="true" slot="icon-right"></i>
 </image-compare>
 ```
+
+### Events
+
+* `@movment` - when image/handle is dragged left/right
 
 ## License
