--- src/index.js
@@ -1,7 +1,0 @@
-import VueImageCompare from './vue-image-compare';
-
-export default {
-  install(Vue) {
-    Vue.component('image-compare', VueImageCompare);
-  }
-}
